enhance discover event UI - filter + event list UI
