--- frontend/src/components/Discover_Event/FilterSection.css
@@ -1,27 +1,0 @@
-.filter-section {
-  width: 250px; /* Adjust width as needed */
-  padding: 20px;
-  /* background-color: greenyellow; */
-  font-size: medium;
-  font-weight: 700;
-  color: grey;
-}
-
-.filter-section h3 {
-  margin-bottom: 10px;
-}
-
-.filter-section div {
-  margin-bottom: 5px;
-}
-
-.filter-section label {
-  margin-left: 10px;
-}
-
-.filter-section button {
-  margin-top: 10px;
-  cursor: pointer;
-}
-
-/* You can add more styling here according to your design preference */
