+++ frontend/src/components/Discover_Event/event-list/EventList.css
