+++ frontend/src/components/Discover_Event/filter-section/FilterSection.css
@@ -1,0 +1,42 @@
+.filter-section {
+  width: 250px; /* Adjust width as needed */
+  padding: 20px;
+  font-size: medium;
+  color: #000;
+}
+
+.filter-section h3 {
+  margin-bottom: 10px;
+  color: #000;
+}
+.filter-section p {
+  font-size: 35px;
+  font-weight: 700;
+}
+
+.filter-section div {
+  margin-bottom: 5px;
+  font-weight: 500;
+}
+
+.filter-section label,
+div {
+  margin-left: 10px;
+  font-weight: 500;
+}
+
+.filter-section button {
+  margin-top: 10px;
+  cursor: pointer;
+  font-weight: 500;
+}
+
+.calendar-container {
+  position: absolute; /* or fixed */
+  top: 30px; /* Adjust this value as needed */
+  left: 0;
+  z-index: 999; /* Ensure the calendar appears above other content */
+  background-color: white; /* Optional: Set background color */
+  border: 1px solid black; /* Optional: Add border */
+  padding: 10px; /* Optional: Add padding */
+}
